requirements.txt
flask
bcrypt
PyJWT
mysql-connector-python
